shum-trading-desk/requirements.txt
pyyaml
jsonschema
pytest
openai
